% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_as_keq.R
\name{chem_as_keq}
\alias{chem_as_keq}
\title{Calculate equilibrium constant through Gibbs free energy and temperature}
\usage{
chem_as_keq(x, y)
}
\arguments{
\item{x}{Gibbs free energy of the reaction}

\item{y}{Temperature of reaction.}
}
\value{
A numeric vector.
}
\description{
Calculate equilibrium constant through Gibbs free energy and temperature
}
\examples{
\dontrun{
k<-chem_as_keq(-200*1000,273.15)
}
}
