% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_as_pkeq.R
\name{chem_as_pkeq}
\alias{chem_as_pkeq}
\title{Calculate equilibrium constant(under log10()) through Gibbs free energy and temperature}
\usage{
chem_as_pkeq(G, T)
}
\arguments{
\item{x}{Gibbs free energy of the reaction}

\item{y}{Temperature of reaction.}
}
\value{
A numeric vector.
}
\description{
Calculate equilibrium constant(under log10()) through Gibbs free energy and temperature
}
\examples{
chem_as_pkeq(-200*1000,273.15)
}
