% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_is_spon.R
\name{chem_is_spon}
\alias{chem_is_spon}
\title{Check if the chemical reaction is spontaneous through enthalpy, entropy and temperature}
\usage{
chem_is_spon(H, S, T)
}
\arguments{
\item{x}{Enthalpy of the reaction}

\item{y}{Entropy of the reaction}

\item{z}{Temperature of reaction.}
}
\value{
A logical vector
}
\description{
Check if the chemical reaction is spontaneous through enthalpy, entropy and temperature
}
\examples{
chem_is_spon(-200*1000,200,273.15)
}
