% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsate.R
\name{gibbsate}
\alias{gibbsate}
\title{Calculate Gibbs free energy through enthalpy, entropy and temperature}
\usage{
gibbsate(x, y, z)
}
\arguments{
\item{x}{Enthalpy of the reaction}

\item{y}{Entropy of the reaction}

\item{z}{Temperature of reaction.}
}
\value{
A numeric vector.
}
\description{
Calculate Gibbs free energy through enthalpy, entropy and temperature
}
\examples{
gibbsate(-200*1000,273.15)
}
